forcing `HTMLCollection` returned from `children` property of an element to be an array.
this handles a weird edge case found on ie11 on win10.
the f12 developer tools will hijack js behavior and make `HTMLCollection`
have associative indexes

ms__id4200
length
item
namedItem

instead of numeric indexes

0
length
item
namedItem

this is necessary for the `each()` function to iterate over
the collection as an array with integer indexes as opposed to
iterating over it as an object, writing the collection's properties
to the DOMQuery.